Update ma_statement.c
Here Stmt->ParamCount is a 2 byte integer and its address is passed as second parameter to mysql_stmt_attr_set(), which is expecting void pointer as its second arguement. That void pointer is upcasted to 4 byte integer at https://github.com/MariaDB/mariadb-connector-c/blob/b00cdcd24f1b191ebfa58f4ff205884a22d8884a/libmariadb/mariadb_stmt.c#L1072. This upcasting works on little endian systems, but on big endian, this changes the value being pointed to. So to fix this, we pass a the address of a 4 byte integer pCount which points to the same value as that pointed by the 2 byte integer Stmt->ParamCount.