Update basic.c
At line 95 and 102, we are passing the address of variable 'value', which is a 4 byte integer, as the 4th argument to function SQLGetData(). This address is used to store the value fetched. But the type of value fetched is SQL_C_USHORT which is a 2 byte integer. A 2 byte integer being read as a 4 byte integer gives the same value on x86 but not on s390x. So we need to change the type of variable 'value' to SQLSMALLINT which is a 2 byte integer.